Trying to receive anything from STM, timeout exception

+++ DataAcquisition/DataContext.cs
@@ -11,7 +11,7 @@ namespace DataAcquisition
         public static string Port { get; set; }
         public static int BufferSize { get; set; }
         public static int HowManyBuffers { get; set; }
-        public static int Frequency { get; set; }
+        public static double Frequency { get; set; }
         public static int HowManyADC { get; set; }
 
     }
